requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
# pydantic==2.5.0
# pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
#sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
scikit-learn==1.3.2
numpy==1.24.3
transformers==4.40.0
torch==2.4.0
google-genai
openai
mistralai
deepseek